Changes needed for IMP-1116.
The IMP-1116 patch changes some of the codec interfaces to int64_t (from int).
This change in Impala-lzo is needed for the code to compile.

(cherry picked from commit 10482edf268b9e57a420523c610a3ce22b48ddc2)

+++ hdfs-lzo-text-scanner.cc
@@ -381,7 +381,7 @@ Status HdfsLzoTextScanner::Checksum(LzoChecksum type, const string& source,
 
 Status HdfsLzoTextScanner::ReadHeader() {
   uint8_t* magic;
-  int bytes_read;
+  int64_t bytes_read;
   Status status;
   // Read the header in. HEADER_SIZE over estimates the maximum header.
   RETURN_IF_FALSE(stream_->GetBytes(HEADER_SIZE, &magic, &bytes_read, &status));
@@ -538,7 +538,7 @@ Status HdfsLzoTextScanner::ReadAndDecompressData() {
 
   // Read in the compressed data
   uint8_t* compressed_data;
-  int bytes_read;
+  int64_t bytes_read;
   RETURN_IF_FALSE(
       stream_->GetBytes(compressed_len, &compressed_data, &bytes_read, &status));
   if (bytes_read == 0) {
